Hook up user interface for tool installation
Add a menu item to bring up the installer dialog, set up the bindings and
connections for initiating installation in response to button clicks, show a
progress indicator and disable the buttons when an install is in action.

This is only the user interface; the back end has not yet been put in place.

Signed-off-by: Wincent Colaiuta <win@wincent.com>